requirements.txt
# RAG Pipeline
chromadb>=0.4.22
sentence-transformers>=2.2.2
openai>=1.0.0
python-dotenv>=1.0.0

# Web UI
gradio>=4.0.0

# Messenger Bot
flask>=3.0.0
requests>=2.31.0
